+++ meteor-components-ioc-plugin.js
@@ -4,6 +4,20 @@ if (Meteor.isClient) {
     return min + Math.random() * (max - min);
   }
 
+  ComponentRootIoc.factory('TestMixin', () => {
+    return {
+      name: 'Test',
+
+      configure(value) {
+        console.log('mixin configured with:', value);
+      },
+
+      ready() {
+        console.log('mixin is ready');
+      }
+    };
+  })
+
   function randomString(len=15, wordLen=5) {
     let letters = [];
     let w = 0;
@@ -48,6 +62,17 @@ if (Meteor.isClient) {
 
     constructor(items) {
       this._items = items;
+    }
+
+    mixins() {
+      return [
+        {
+          name: 'Draggable',
+          configure(value) {
+            console.log('Draggable configured with:', value);
+          }
+        }
+      ];
     }
 
     helpers() {