+++ packages/meteor-components-ioc-plugin/meteor-components-ioc-plugin.js
@@ -1,4 +1,4 @@
-/*global ComponentRootIoc, Package, ReactiveObj, EJSON*/
+/*global ComponentRootIoc, Template, Package, EJSON*/
 function install(Component, ComponentUtil, IocContainer) {
   ComponentRootIoc = new IocContainer();
 
@@ -22,7 +22,7 @@ function install(Component, ComponentUtil, IocContainer) {
             initializing(component) {
               component.name = componentName;
               component.templateInstance = templateInstance;
-              overrideMixins(component, ioc);
+              initializeMixins(component, templateInstance, ioc);
               Component.trigger('initializing', component, templateInstance);
             }
           }
@@ -65,54 +65,97 @@ function getNearestIocContainer(view, defaultIoc) {
   return view ? (view.ioc || defaultIoc) : defaultIoc;
 }
 
-function overrideMixins(component, ioc) {
+function initializeMixins(component, templateInstance, ioc) {
   // Override mixins so that dependencies are injected and
   // each mixin is managed.
   let mixinsFn = component.mixins;
-  let mixinsInstalled = true;
+  let mixins = [];
 
   if (typeof mixinsFn === 'function') {
-    component.mixins = function () {
-      let mixins = mixinsFn.call(this) || [];
-      if (mixinsInstalled) return mixins;
-      mixinsInstalled = true;
+    // Instantiate the mixins right away so that they will be
+    // tracked as dependencies for the component instance.
+    mixins = mixinsFn.call(this) || [];
+    for (let k = 0, len = mixins.length; k < len; k += 1) {
+      let m = mixins[k];
 
-      return mixins.map(function (m, index) {
-        let options = {
-          transient: false,
-          concerns: {
-            initializing(mixin) {
-              // Ensure that the destroy method only gets called once.
-              // This has to be done because the mixin plugin that is
-              // defined in meteor-components will call mixin.destroy()
-              // just before the component is destroyed and the IOC
-              // container will also call destroy when the IOC container
-              // is destroyed.
-              let destroy = mixin.destroy;
-              let destroyed = false;
-              if (typeof destroy === 'function') {
-                mixin.destroy = function () {
-                  if (destroyed) return;
-                  destroyed = true;
-                  destroy.call(this);
-                };
+      if (typeof m === 'function') {
+        m = ioc.inject(m);
+        // Update the reference in the array.
+        mixins[k] = m;
+      }
+
+      // If the mixin is named then we mark it as being registered
+      // so it doesn't get added twice to the final mixins list and
+      // we also register it as a constant in the IOC container so
+      // that it can be configured by an attached mixin property.
+      if (typeof m.name === 'string' && m.name) {
+        m.$registered = true;
+        ioc.constant(m.name + 'Mixin', m);
+      }
+
+      let destroy = m.destroy;
+      let destroyed = false;
+      if (typeof destroy === 'function') {
+        m.destroy = function () {
+          if (destroyed) return;
+          destroyed = true;
+          destroy.call(this);
+        };
+      }
+    }
+  }
+
+  let attachedMixins = initializeAttachedMixins(
+    component,
+    templateInstance,
+    ioc
+  );
+
+  mixins = mixins.concat(attachedMixins);
+  component.mixins = () => mixins;
+}
+
+function initializeAttachedMixins(component, templateInstance, ioc) {
+  let data = templateInstance.data || {};
+  let prefix = 'Mixin.';
+  let mixins = [];
+
+  for (let key in data) {
+    if (key.indexOf(prefix) === 0 && key.length > prefix.length) {
+      // Attempt to lookup 'NameMixin' as a service.
+      let mixinName = key.substr(prefix.length) + 'Mixin';
+      if (ioc.canResolve(mixinName)) {
+
+        let mixin = ioc.resolve(mixinName);
+        if (typeof mixin.configure === 'function') {
+
+          // Only call configure() when the key has changed.
+          let setting = {};
+          templateInstance.autorun(function (c) {
+            let data = Template.currentData();
+            let newSetting = data[key];
+            if (newSetting !== setting) {
+              setting = newSetting;
+              if (setting !== undefined) {
+                mixin.configure(newSetting);
               }
             }
-          }
-        };
-
-        if (typeof m !== 'function') {
-          ioc.factory(`mixin_${index}`, m, options);
-        } else {
-          ioc.factory(`mixin_${index}`, () => Object.create(m), options);
+          });
         }
-
-        return function () {
-          return ioc.resolve(`mixin_${index}`);
-        };
-      });
-    };
+        // If the mixin has already been registered in the IOC container
+        // then we don't add it to the mixins list since it will already
+        // have been added and tracked. This mixins list is for newly
+        // instantiated mixins.
+        // NOTE: This is an optimization so that we don't have to filter
+        // out registered mixins.
+        if (!!mixin.$registered) mixins.push(mixin);
+      } else {
+        throw new Error('Attached mixin (' + mixinName + ') not found.');
+      }
+    }
   }
+
+  return mixins;
 }
 
 // If the weak dependencies exist then we install the plugin.